+++ modules/tripexplorer/SubwayTripGraphs.tsx
@@ -30,6 +30,7 @@ export const SubwayTripGraphs: React.FC<SubwayTripGraphsProps> = ({
   line,
 }) => {
   const [peakTime, setPeakTime] = React.useState<'weekday' | 'weekend'>('weekday');
+
   const { traveltimes, headways, dwells } = useTripExplorerQueries(
     'subway',
     parameters,
@@ -81,13 +82,7 @@ export const SubwayTripGraphs: React.FC<SubwayTripGraphsProps> = ({
             />
           </WidgetDiv>
           <WidgetDiv className="flex flex-col justify-center">
-            <WidgetTitle
-              title="Travel Times"
-              subtitle="By Hour"
-              location={location}
-              line={line}
-              both
-            />
+            <WidgetTitle title="Travel Times by Hour" location={location} line={line} both />
             <AggregateChartWrapper
               query={traveltimes}
               toStation={toStation}
